dtos

+++ eshopping-backend/src/main/java/com/eshopping/eshopping_backend/model/Product.java
@@ -16,11 +16,11 @@ import lombok.Setter;
 public class Product {
      @Id
      @GeneratedValue(strategy = GenerationType.IDENTITY)
-     private long id ;
+     private Long id ;
 
      @Column
      private String name ;
      @Column
-     private float price ;
+     private double price ;
 
 }
